Anonymous Nuker
This source can attack ANY servers with only 1 command and 1 message sent to chat!

Use !nuke For All channels & roles gets deleted and members gets banned except you.

Type in spam For A Loads of Channels Named F#cked-By-Anonymous. If the spam runs for 30s, it would take too 1k+ Tags.

Use disable for turning off the bot and end the process(for attacking again, turn on the bot)